Add ESSHelper.readCompound(InputStream)
